+++ R/preprocessing.R
@@ -188,7 +188,11 @@ fps_format_factors <- function(.data,
                   fps_slr_name = {{farm_size_col}},
                   fps_robust = {{farm_type_col}},
                   post_strat = {{strat_col}}) %>%
-    dplyr::mutate(fps_gor = gsub(" \\(England\\)", "", fps_gor, ignore.case = TRUE),
+    dplyr::mutate(fps_gor = trimws(fps_gor),
+                  fps_gor = gsub(" \\(England\\)", "", fps_gor, ignore.case = TRUE),
+                  fps_gor = ifelse(grepl("London", fps_gor, ignore.case = TRUE), "South East including London", fps_gor),
+                  fps_gor = ifelse(grepl("South East", fps_gor, ignore.case = TRUE), "South East including London", fps_gor),
+                  fps_gor = ifelse(grepl("^East$", fps_gor, ignore.case = TRUE), "East of England", fps_gor),
                   fps_gor = ifelse(grepl("North West", fps_gor, ignore.case = TRUE), "North West and Merseyside", fps_gor),
 
                   fps_slr_name = snakecase::to_sentence_case(fps_slr_name),
@@ -196,6 +200,16 @@ fps_format_factors <- function(.data,
                   fps_robust = ifelse(grepl("pig|poultry", fps_robust, ignore.case = TRUE), "Pigs and poultry", fps_robust),
 
                   post_strat = as.character(post_strat))
+
+  if(any(is.na(.data$fps_gor))) {
+    cli::cli_alert_danger("There are `NA` records in the data for `fps_gor`. Check the factor levels in the data.")
+  }
+  if(any(is.na(.data$fps_slr_name))) {
+    cli::cli_alert_danger("There are `NA` records in the data for `fps_slr_name`. Check the factor levels in the data.")
+  }
+  if(any(is.na(.data$fps_robust))) {
+    cli::cli_alert_danger("There are `NA` records in the data for `fps_robust`. Check the factor levels in the data.")
+  }
 
   return(.data)